net: phy: mxl-gpy: add LED dimming support
Some PHYs support LED dimming. The use case is a router that dims LEDs
at night. As of now, only Maxlinear PHYs support dimming. In the future,
support may be extended to other PHYs such as the Realtek RTL8221B.

PHYs from different manufacturers support a different number of
brightness levels, so it was necessary to extend the API with the
led_max_brightness() function. If this function is omitted, a default
value is used, assuming that only two levels are supported.

Signed-off-by: Aleksander Jan Bajkowski <olek2@wp.pl>
Reviewed-by: Daniel Golle <daniel@makrotopia.org>