phy: realtek: test cable diagnostic on RTL8221B
The RTL8224 can detect open pairs and short types (in same pair or some
other pair). The distance to this problem can be estimated. This is done
for each of the 4 pairs separately.

It is not meant to be run while there is an active link partner because
this interferes with the active test pulses.

Signed-off-by: Aleksander Jan Bajkowski <olek2@wp.pl>