net: phy: mediatek: make PHY LEDs optional on AN758x
On some devices, the LED pins on the built-in PHY are not used. On other
devices, they are used by other peripherals. An example is the Gemtek W1700,
which uses some of these pins as interrupts from an external PHY. Currently,
the driver prints a message about a missing pinctr, which confuses users. Do
not print a message about a missing pinctrl if the LEDs are not used.

Signed-off-by: Aleksander Jan Bajkowski <olek2@wp.pl>